+++ src/main/java/com/unrealdinnerbone/petrock/entity/PREntitys.java
@@ -1,0 +1,18 @@
+package com.unrealdinnerbone.petrock.entity;
+
+import com.unrealdinnerbone.petrock.libs.util.EntityHelper;
+
+/**
+ * Created by manmaed on 06/11/2015.
+ */
+public class PREntitys {
+
+    public static void load() {
+
+        /* EntityHelper.registerEntity(EntityClassName.class, "Name",  bkEggColor, dotEggColor); */
+        EntityHelper.registerEntity(EntityPetRockFollowing.class, "PetRockFollowing", 0x292929, 0xd6d6d6);
+        EntityHelper.registerEntity(EntityPetRockStay.class, "PetRockStay", 0xd6d6d6, 0x292929);
+
+    }
+
+}
