+++ src/main/java/com/unrealdinnerbone/petrock/libs/Reference.java
@@ -1,0 +1,12 @@
+package com.unrealdinnerbone.petrock.libs;
+
+public class Reference {
+    //All the Strings
+    public static final String MOD_ID = "petrock";
+    public static final String MOD_NAME = "PetRock";
+    public static final String VERSION = "0.1.0";
+    public static final String CLIENT_PROXY = "com.unrealdinnerbone.petrock.proxy.ClientProxy";
+    public static final String PROXY_COMMON = "com.unrealdinnerbone.petrock.proxy.CommonProxy";
+    public static final String RE_PREFIX = MOD_ID.toLowerCase() + ":";
+
+}
