+++ src/main/java/com/unrealdinnerbone/petrock/proxy/CommonProxy.java
@@ -1,0 +1,20 @@
+package com.unrealdinnerbone.petrock.proxy;
+
+/**
+ * Created by manmaed on 05/11/2015.
+ */
+public class CommonProxy {
+
+    public void preInit() {
+
+    }
+
+    public void init() {
+
+    }
+
+    public void renderInformation() {
+
+    }
+
+}
