--- src/main/java/manmaed/petrock/blocks/PRBlocks.java
@@ -1,17 +1,0 @@
-package manmaed.petrock.blocks;
-
-import net.minecraft.block.Block;
-
-/**
- * Created by manmaed on 10/11/2015.
- */
-public class PRBlocks {
-
-    public static Block blockpetrock;
-
-    public static void load() {
-
-        blockpetrock = new BlockPetRock();
-
-    }
-}
