--- src/main/java/manmaed/petrock/libs/Reference.java
@@ -1,14 +1,0 @@
-package manmaed.petrock.libs;
-
-/**
- * Created by manmaed on 05/11/2015.
- */
-public class Reference {
-    //All the Strings
-    public static final String MOD_ID = "petrock";
-    public static final String MOD_NAME = "PetRock";
-    public static final String VERSION = "0.1.0";
-    public static final String CLIENT_PROXY = "manmaed.petrock.proxy.ClientProxy";
-    public static final String PROXY_COMMON = "manmaed.petrock.proxy.CommonProxy";
-
-}
