--- src/main/java/manmaed/petrock/libs/util/EntityHelper.java
@@ -1,20 +1,0 @@
-package manmaed.petrock.libs.util;
-
-import cpw.mods.fml.common.registry.EntityRegistry;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityList;
-
-/**
- * Created by manmaed on 06/11/2015.
- */
-public class EntityHelper {
-
-    public static void registerEntity(Class<? extends Entity> entityClass, String entityName, int bkEggColor, int fgEggColor) {
-        int id = EntityRegistry.findGlobalUniqueEntityId();
-
-        EntityRegistry.registerGlobalEntityID(entityClass, entityName, id);
-        EntityList.entityEggs.put(Integer.valueOf(id), new EntityList.EntityEggInfo(id, bkEggColor, fgEggColor));
-
-    }
-
-}
