--- src/main/java/manmaed/petrock/proxy/ClientProxy.java
@@ -1,29 +1,0 @@
-package manmaed.petrock.proxy;
-
-import cpw.mods.fml.client.registry.RenderingRegistry;
-import manmaed.petrock.client.render.entity.RenderPetRockFollowing;
-import manmaed.petrock.client.render.entity.RenderPetRockStay;
-import manmaed.petrock.client.render.model.ModelPetRockFollowing;
-import manmaed.petrock.client.render.model.ModelPetRockStay;
-import manmaed.petrock.entity.EntityPetRockFollowing;
-import manmaed.petrock.entity.EntityPetRockStay;
-
-/**
- * Created by manmaed on 05/11/2015.
- */
-public class ClientProxy extends CommonProxy {
-
-    public void preInit() {
-
-    }
-    public void init() {
-
-    }
-
-    public void renderInformation() {
-
-        RenderingRegistry.registerEntityRenderingHandler(EntityPetRockFollowing.class, new RenderPetRockFollowing(new ModelPetRockFollowing(0), 0.5f));
-        RenderingRegistry.registerEntityRenderingHandler(EntityPetRockStay.class, new RenderPetRockStay(new ModelPetRockStay(0), 0.5f));
-
-    }
-}
