--- src/main/java/manmaed/petrock/proxy/CommonProxy.java
@@ -1,20 +1,0 @@
-package manmaed.petrock.proxy;
-
-/**
- * Created by manmaed on 05/11/2015.
- */
-public class CommonProxy {
-
-    public void preInit() {
-
-    }
-
-    public void init() {
-
-    }
-
-    public void renderInformation() {
-
-    }
-
-}
