+++ src/main/resources/assets/petrock/models/block/BlockPetRock.json
@@ -1,0 +1,3 @@
+{
+    "parent": "block/cube_all"
+}
