+++ src/main/resources/mcmod.info
@@ -7,8 +7,8 @@
   "mcversion": "1.7.10",
   "url": "",
   "updateUrl": "",
-  "authorList": ["manmaed"],
-  "credits": "Yunus1903, KakesRevenge, TheKayneGame",
+  "authorList": ["unrealdinnerbone"],
+  "credits": "Yunus1903, KakesRevenge, TheKayneGame, manmaed",
   "logoFile": "",
   "screenshots": [],
   "dependencies": []
